Improved documentation and new atmospheric pressure test.
Updated documentation within the methods and function calls.

Added a new test to use MetPy pressure derived from altitude
to check if values are within expected range.

Fixed issue with writing to file in CF format and reading file
back into Dataset. There was an issue with the flag_* variable
attributes not being in a list or array as expected when only one
value exists.

Added additional tests for the new atmospheric pressure testing
function.