Adding keep_variables to reader
The reader allows the use of drop_variables to not read in variables
but this requires the knowledge of the variable names to drop. This
feature allows for setting the variables to keep and it will figure the
names of the variables to add to drop_variables to not read in.