Improved documentation and new atmospheric pressure test. (#466)
* Improved documentation and new atmospheric pressure test.

Updated documentation within the methods and function calls.

Added a new test to use MetPy pressure derived from altitude
to check if values are within expected range.

Fixed issue with writing to file in CF format and reading file
back into Dataset. There was an issue with the flag_* variable
attributes not being in a list or array as expected when only one
value exists.

Added additional tests for the new atmospheric pressure testing
function.

* Removing deletion of object after use in case this is causing the issue when testing on Windows OS.

* Fixing permissions issue

When testing in windows there is a strange issue with not able to unlink
a file. Testing if this change fixes that issue.