Fixing QC test removal when flag_masks is numpy array.
When the test is removed, the test number in flag_masks is removed using
list syntax. Depending on how the netCDF data is written to file the
flag_masks can be set as list of integers or numpy array of numpy
integers. Needed to change method of removing test based on data type.