Documentation update (#465)
* Cleaning up documentation

Noticed the documentation was not consistent and needed to be updated.
Also noticed qc variable name keyword was never used. Removed that
keyword option and added another option to apply cleanup method.
Left the default to cleanup as before.

* Documentation update and improved parse_bits

Updated documentation to be consistent and more accurate. Noticed
not every method/function had example. Added example were lacking.

Discovered a better way to parse bits using numpy instead of
a while loop. Added documentation describing how it works. Left the old
method in for now as a reference. The new method works up to 64 bit
integers in numpy only, while old method theortically would work
with larger integer values. Since can't use larger integer values
with numpy not concerned.

Also updated unset_bit to be consistent with set_bit in the bit
minipulation syntax.

* Removing keyword

Removing call to keyword that actually was not implemented
in the function and is now not a callable keyword.