QC Configuration file (#468)
* Fixing QC test removal when flag_masks is numpy array.

When the test is removed, the test number in flag_masks is removed using
list syntax. Depending on how the netCDF data is written to file the
flag_masks can be set as list of integers or numpy array of numpy
integers. Needed to change method of removing test based on data type.

* Feature: Adding ability to apply a YAML configuration file of QC

This new feature allows storing tests and times in a YAML file that can
be applied to the Dataset as new tests in the ancillary quality control
variable. Will read the YAML file and add a new test for each test
listed using the assessment provided. Will create a new QC variable if
one does not exists. Added example for how to use the three main QC
parts, embedded QC, DQRs and YAML tests.

* Adding a file level doc string.

* Fixing username and token strings to pass builing documentation.

* Updated username and token to match from plot_multiple_dataset.py

* Adding missing import statement

* Made minor change to force running again to see if the code will pass if run again.

* ENH: Adding to init files

* ENH: Making some updates to correct typos and ensure only one MFRSR.  Also adding in check on the example to ensure it's getting a token

* Changing the name of the supplemental QC file and functions to better
describe the funciton and module.

* Updating the documentation.

Co-authored-by: Adam Theisen <atheisen@anl.gov>